+++ lib/an_lifecycle_cancellable.dart
@@ -8,4 +8,5 @@ export 'lifecycle/lifecycle_async.dart';
 export 'lifecycle/lifecycle_data.dart';
 export 'lifecycle/lifecycle_effect.dart';
 export 'lifecycle/lifecycle_ext.dart';
+export 'lifecycle/lifecycle_context_effect.dart';
 export 'listenable/change_notifier_ext.dart';